Hook replacement to original place
