+++ web/src/components/hostonboarding/views/PhotosView.js
@@ -1,6 +1,6 @@
 import { useParams } from "react-router-dom";
 import ImagePreview from "../components/ImagePreview";
-import usePhotos from "./usePhotos";
+import usePhotos from "../hooks/usePhotos";
 import Button from "../components/button";
 import { useRef, useState } from "react";
 import "./PhotoVieuw.css";
